fix test

+++ x/lockup/keeper/lock_test.go
@@ -671,7 +671,7 @@ func (suite *KeeperTestSuite) TestEditLockup() {
 	// extending lock with same duration should fail
 	err = suite.App.LockupKeeper.ExtendLockup(suite.Ctx, *lock, time.Second)
 	suite.Require().Error(err)
-	
+
 	// duration increase should success
 	err = suite.App.LockupKeeper.ExtendLockup(suite.Ctx, *lock, time.Second*2)
 	suite.Require().NoError(err)
@@ -679,9 +679,9 @@ func (suite *KeeperTestSuite) TestEditLockup() {
 	// check queries
 	lock, _ = suite.App.LockupKeeper.GetLockByID(suite.Ctx, lock.ID)
 	suite.Require().Equal(lock.Duration, time.Second*2)
-	suite.Require().Equal(1, lock.ID)
+	suite.Require().Equal(uint64(1), lock.ID)
 	suite.Require().Equal(coins, lock.Coins)
-	
+
 	locks = suite.App.LockupKeeper.GetLocksLongerThanDurationDenom(suite.Ctx, "stake", time.Second)
 	suite.Require().Equal(len(locks), 1)
